requirements.txt
torch>=2.10.0
matplotlib>=3.10.8
hydra-core>=1.3.2
omegaconf>=2.3.0
